fix: move type
