+++ src/hooks/queries/useBookInfo/dataTransforms.ts
@@ -1,19 +1,11 @@
-import type { BooksResponse } from './types'
-
-export type BookInfo = {
-  title: string
-  authors: string[]
-  description: string
-  image: string
-  publishedDate: string
-}
+import type { BooksResponse, BookInfo } from './types'
 
 /**
  * Methods to modify the data from the requests into the desired format
  * @returns An object with methods to transform the data
  */
 export const useBookInfoTransforms = () => {
-  const mapBookInfo = (data: BooksResponse) => {
+  const mapBookInfo = (data: BooksResponse): BookInfo => {
     const bookInfo = data.items[0]
     return {
       title: bookInfo.volumeInfo.title,
